Cleanup readline parser validator implementation
Move the `ParserValidator` type to its own module in `artichoke::repl::parser`.
This allows the `cfg`s to be added in only one spot and to simplify the
type names.

Additionally, using the `Editor::get_helper` API, cloning the parser
validator and wrapping in `Arc` is no longer necessary, although drop
order becomes more important to the structure of the code.

Additionally, interpreter init and preamble printing is split out into a
separate `init` routine and readline editor setup is moved into the
`run` function. This means `artichoke::repl::run` imperatively prepares
the REPL environment. Then `artichoke::repl::repl_loop` only implements
the REPL loop itself.

Additionally, `exit` is now added to the readline history.

All internal functions to the `artichoke::repl` module are now
documented.
